restructure results folder
