+++ frontend/src/features/event/results/attendees/panel-header.tsx
