+++ frontend/src/features/event/results/attendees/participant-chip.tsx
