+++ frontend/src/features/event/results/attendees/remove-participant.tsx
