+++ frontend/src/features/event/results/components/banners.tsx
