+++ frontend/src/features/event/results/components/display-settings.tsx
